:white_check_mark: [#297] Test StufBg get_children
